requirements.txt
amqp==5.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
constantly==23.10.4
coverage==7.6.7
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect==1.2.0
daphne==4.1.2
defusedxml==0.7.1
Deprecated==1.2.15
dj-database-url==2.3.0
Django>=3.2,<4.0
django-dotenv==1.4.2
django-widget-tweaks==1.5.0
fake-useragent==1.5.1
feedparser==6.0.11
gevent==22.10.2
gunicorn
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.1.3
packaging==24.2
parse==1.20.2
pillow==11.0.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==11.1.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.5.0
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
stripe==11.2.0
tqdm==4.67.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
websockets==10.4
whitenoise
wrapt==1.16.0
yahoo-finance==1.4.0
zipp==3.21.0
zope.interface==7.1.1
psycopg2-binary
